requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0
pyyaml>=6.0.0
pydantic>=2.0.0
